requirements.txt
pystan==2.19.1.1
fbprophet
numpy==1.19.0
torch==1.6.0
pandas==1.1.0
sklearn==0.0
scikit-learn==0.23.2
neuralprophet==0.2.7
matplotlib==3.2.2
pytorch-lightning==1.2.1
holidays==0.10.2
scipy==1.5.0
tqdm==4.46.1
tensorboard==2.4.0
u8darts[all]==0.6.1
plotly==4.14.3
wandb==0.10.21
cufflinks
jupyterlab
notebook